fix(core): Add checks for the existence of window prior to attempting a polyfill to prevent server c
Running a project that includes videogular2 and server side rendering will crash with
ReferenceError: window is not defined on server start.  This is because the polyfill attempts to
apply regardless of whether or not you are attempting to render a page with a video.  This change
checks if window is undefined before attempting to access the window variable, preventing this
server crash.