fix(vgAPI): Fix syntax error in getMasterMedia and isMasterDefined functions
getMasterMedia and isMasterDefined both were checking for vgMedia (an IMediaElement) being equal to
the boolean value 'true' rather than the boolean value vgMaster.  This means that deployments that
include a [vgMaster]="true" definition always fail the test which makes getMasterMedia return
defaultMedia, equal to the first media object in the media collection.  This means that api both
reports values in $$getAllProperties incorrectly

#729